fix: bootstrap AUC CI silently overflowed int32 on Windows
The "impossible" bootstrap CI values from today's training runs
(xgb_ci95_lo/hi like [15.84, 18.49] instead of a sane range around the
0.83 point estimate) were real, reproducible output, not a copy/paste
artifact as first suspected -- confirmed once a directly-formatted
logger.info() line (no DataFrame rendering involved) showed the same
values.

Root cause: pandas' bare .astype(int) resolves to the platform's native
int width -- int64 on Linux/Mac, but int32 on Windows. _batched_auc's
n_pos * n_neg (and n_pos * (n_pos + 1)) silently overflows a 32-bit
accumulator once n_rows reaches the tens of thousands (a real run at
n=225,940 had a true product of ~9.2B against an int32 max of ~2.15B),
producing a wrapped, sometimes-negative denominator with no error raised.
Reproduced exactly: forcing int32 accumulation on a synthetic 100k-row
case turns a true product of 2,500,000,000 into -1,794,967,296.

Fixed by explicitly forcing dtype=np.int64 both where _batched_auc sums
n_pos and at the upstream cmp_df["y_xgb"]/["y_lgb"] cast, rather than
relying on either the caller's dtype or numpy's platform-dependent
promotion. Point estimates (roc_auc_score-based) were never affected --
only the bootstrap-derived CI bounds, which is exactly the pattern
observed across every affected run this session.

Added a regression test constructing int32 input at overflow-triggering
scale, plus a source-inspection test pinning the explicit dtype=np.int64 --
the former alone can't catch a reversion on 64-bit Linux CI, since ambient
numpy .sum() promotion already lands on int64 there, masking a bug that
only manifests on Windows. Neither of the existing 488 tests caught this,
since rng.integers() defaults to int64 on every platform and the existing
scale (n_rows=40) was far too small to overflow regardless.

+++ pd_model/modeling/calibration.py
@@ -411,7 +411,17 @@ def _batched_auc(y_b: np.ndarray, score_b: np.ndarray) -> np.ndarray:
     see test_calibration.py).
     """
     ranks = rankdata(score_b, method="average", axis=1)
-    n_pos = y_b.sum(axis=1)
+    # dtype=np.int64 is required, not incidental -- numpy's .sum() promotes
+    # an integer accumulator only to the platform's native int width, which
+    # is int64 on Linux/Mac but int32 on Windows. n_pos * n_neg (and
+    # n_pos * (n_pos + 1)) silently overflows a 32-bit accumulator once
+    # n_rows reaches the tens of thousands (e.g. n_pos=n_neg=50,000 gives
+    # a true product of 2.5B against an int32 max of ~2.15B), producing a
+    # wrapped, sometimes-negative denominator and wildly wrong AUC/CI
+    # values with no error raised -- confirmed to reproduce exactly this
+    # way on a real bootstrap run. int64 has no realistic overflow risk at
+    # any dataset size this code will see.
+    n_pos = y_b.sum(axis=1, dtype=np.int64)
     n_neg = y_b.shape[1] - n_pos
     sum_pos_ranks = (ranks * y_b).sum(axis=1)
     with np.errstate(invalid="ignore", divide="ignore"):
@@ -482,8 +492,12 @@ def run_bootstrap_comparison(
         ).dropna()
         align_mode = "index_intersection"
 
-    cmp_df["y_xgb"] = cmp_df["y_xgb"].astype(int)
-    cmp_df["y_lgb"] = cmp_df["y_lgb"].astype(int)
+    # int64, not bare int -- pandas' .astype(int) resolves to the platform's
+    # native int width (int32 on Windows, int64 on Linux/Mac). _batched_auc
+    # now forces its own int64 accumulator regardless, but fixing the dtype
+    # at its source here too avoids depending on that alone.
+    cmp_df["y_xgb"] = cmp_df["y_xgb"].astype(np.int64)
+    cmp_df["y_lgb"] = cmp_df["y_lgb"].astype(np.int64)
     y_mismatch = int((cmp_df["y_xgb"] != cmp_df["y_lgb"]).sum())
     cmp_df = cmp_df.drop(columns=["y_lgb"]).rename(columns={"y_xgb": "y"})
